Configure the scheduled lambda function
Configure the lambda function, based on the code of [1]. It will
start the code that runs the ECS task. It passes all the required
variables

Additionally the variables that are passed to the container, although
it is not really required.

The lambda function will have a role that allows:
     1. Write the logs in cloudwatch
     2. Start the ECS task and assume the role of the ECS task

We also configure a cloudwatch even with a cron expression that would
trigger the lambda function.

Finally we get as parameters:
 - cluster name
 - command to run
 - cron expression